requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-community>=0.0.200
huggingface-hub>=0.16.4
transformers>=4.40.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
sentence-transformers
streamlit